+++ __fxstat.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -16,7 +16,7 @@ extern int __fxstatat(int, int, const char *, struct stat *, int);
 
 int __fxstat(int ver, int dfd, struct stat *statbuf)
 {
-	__debug("%s(dfd: %i, ...)\n", __func__, dfd);
+	__debug("%s(dfd: %i <-> '%s', ...)\n", __func__, dfd, __fpath(dfd));
 
 	return __fxstatat(ver, dfd, "", statbuf, AT_EMPTY_PATH);
 }