+++ faccessat.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -45,8 +45,8 @@ int faccessat(int dfd, const char *path, int mode, int atflags)
 		return -1;
 	}
 
-	__debug("%s(dfd: %i, path: '%s' -> '%s', mode: 0%03o, atflags: 0x%x)\n",
-		__func__, dfd, path, buf, mode, atflags);
+	__debug("%s(dfd: %i <-> '%s', path: '%s' -> '%s', mode: 0%03o, atflags: 0x%x)\n",
+		__func__, dfd, __fpath(dfd), path, buf, mode, atflags);
 
 	__remove_at_empty_path_if_needed(buf, atflags);
 	return next_faccessat(dfd, buf, mode, atflags);