+++ fstatat64.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -57,8 +57,8 @@ int fstatat64(int dfd, const char *path, struct stat64 *statbuf, int atflags)
 		return -1;
 	}
 
-	__debug("%s(dfd: %i, path: '%s' -> '%s', ..., atflags: 0x%x)\n",
-		__func__, dfd, path, buf, atflags);
+	__debug("%s(dfd: %i <-> %s, path: '%s' -> '%s', ..., atflags: 0x%x)\n",
+		__func__, dfd, __fpath(dfd), path, buf, atflags);
 
 	__remove_at_empty_path_if_needed(buf, atflags);
 	ret = next_fstatat64(dfd, buf, statbuf, atflags);