+++ futimesat.c
@@ -39,7 +39,8 @@ int __futimesat(int dirfd, const char *pathname, const struct timeval times[2])
 
 int futimesat(int dfd, const char *path, const struct timeval times[2])
 {
-	__debug("%s(dfd: %i, path: '%s', ...)\n", __func__, dfd, path);
+	__debug("%s(dfd: %i <-> '%s', path: '%s', ...)\n", __func__, dfd,
+		__fpath(dfd), path);
 
 	return __futimesat(dfd, path, times);
 }