+++ mkdirat.c
@@ -54,8 +54,8 @@ int mkdirat(int dfd, const char *path, mode_t mode)
 	}
 
 	__fwarn_if_insuffisant_user_modeat(dfd, buf, mode, 0);
-	__debug("%s(dfd: %i, path: '%s' -> '%s', mode: 0%03o -> 0%03o)\n",
-		__func__, dfd, path, buf, oldmode, mode);
+	__debug("%s(dfd: %i <-> '%s', path: '%s' -> '%s', mode: 0%03o -> 0%03o)\n",
+		__func__, dfd, __fpath(dfd), path, buf, oldmode, mode);
 
 	ret = next_mkdirat(dfd, buf, mode);
 	__set_mode(buf, oldmode, mode);