+++ openat.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 Gaël PORTAY
+ * Copyright 2021-2023 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -70,8 +70,8 @@ int openat(int dfd, const char *path, int oflags, ...)
 
 	if (oflags & O_CREAT)
 		__fwarn_if_insuffisant_user_modeat(dfd, buf, mode, 0);
-	__debug("%s(dfd: %i, path: '%s' -> '%s', oflags: 0%o, mode: 0%03o -> 0%03o)\n",
-		__func__, dfd, path, buf, oflags, oldmode, mode);
+	__debug("%s(dfd: %i <-> '%s', path: '%s' -> '%s', oflags: 0%o, mode: 0%03o -> 0%03o)\n",
+		__func__, dfd, __fpath(dfd), path, buf, oflags, oldmode, mode);
 
 	ret = next_openat(dfd, buf, oflags, mode);
 	__set_mode(buf, oldmode, mode);