+++ path_resolution.c
@@ -531,7 +531,25 @@ ssize_t fpath(int fd, char *buf, size_t bufsiz)
 	return siz;
 }
 
-static char *__fpath(int fd)
+__attribute__((visibility("hidden")))
+char *__fpath(int fd)
+{
+	const int save_errno = errno;
+	static char buf[PATH_MAX];
+	ssize_t siz;
+
+	*buf = 0;
+	siz = fpath(fd, buf, sizeof(buf));
+	if (siz == -1) {
+		errno = save_errno;
+		return NULL;
+	}
+
+	return buf;
+}
+
+__attribute__((visibility("hidden")))
+char *__fpath2(int fd)
 {
 	const int save_errno = errno;
 	static char buf[PATH_MAX];